pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["alexey <a556077@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
minio = "^7.1.17"
django-minio-storage = "^0.5.7"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.12"
django-redis = "^5.4.0"
psycopg-binary = "^3.1.13"
django-countries = "^7.5.1"
django-filter = "^23.5"
django-admin-sortable2 = "^2.1.10"
django-admin-search = "^0.3.15"
playwright = "^1.43.0"
requests = "^2.31.0"
curlify = "^2.2.1"
pycurl = "^7.45.3"
redis = "^5.0.4"
djangorestframework-simplejwt = "^5.3.1"
stripe = "^9.6.0"
drf-yasg = "^1.21.7"
pyjwt = "^2.8.0"
celery = "^5.4.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
force_single_line = "True"